NAPRAWA BŁEDÓW JAWORSKIEGO FIX

+++ src/main/java/com/example/miniusos/model/DolaczDoPrzedmiotForm.java
@@ -5,7 +5,7 @@ public class DolaczDoPrzedmiotForm {
     private int obecna_ilosc_studentow;
     private Student student;
     private Przedmiot przedmiot;
-    private int id_przycisk;
+    private Long id_przycisk;
 
     public Przedmiot getPrzedmiot() {
         return przedmiot;
@@ -23,11 +23,11 @@ public class DolaczDoPrzedmiotForm {
         this.student = student;
     }
 
-    public int getId_przycisk() {
+    public Long getId_przycisk() {
         return id_przycisk;
     }
 
-    public void setId_przycisk(int id_przycisk) {
+    public void setId_przycisk(Long id_przycisk) {
         this.id_przycisk = id_przycisk;
     }
